I refactored the Gallery.js component into subcomponents named GalleryComponent, ResultSelection & ImageCard (which is further broken down into ImageCardHeader, ImageCardLikes, ImageCardBody, and 'ImagesCardComments').
GalleryComponent: This is the main component that sets up the layout for the gallery. It will function as a container that loops through the data and passes it to individual ImageCard.

ResultSelection component: takes the images and a handleResultsChange function as props to create a dropdown that allows a user to select the number of image cards they want to display on each page.

ImageCard: As the main component responsible for rendering each image card, it's clearly segmented with header and body as separate components. This provides easier management in future.

ImageCardHeader: This component is perfectly embodying the single responsibility principle; it only takes care of displaying the title.

ImageCardBody: It correctly receives the necessary props and diligently hands over the required roles to ImageCardLikes and ImagesCardComments.

ImageCardLikes: This component has been valuable in simplifying the ImageCardBody component, providing the logic and layout for displaying likes and dislikes.